Module pallet_foreign_investments::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

Of

Aux type for configurations that inherents from Config

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Aliases

CollectedInvestment

Maps an investor and their InvestmentId to the collected investment amount, i.e., the payment amount of pool currency burned for the conversion into collected amount of tranche tokens based on the fulfillment price(s).

CollectedRedemption

Maps an investor and their InvestmentId to the collected redemption amount, i.e., the payment amount of tranche tokens burned for the conversion into collected pool currency based on the fulfillment price(s).

InvestmentPaymentCurrency

Maps an investor and their investment id to the foreign payment currency provided on the initial investment increment.

InvestmentState

Maps an investor and their InvestmentId to the corresponding InvestState.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

RedemptionPayoutCurrency

Maps an investor and their investment id to the foreign payout currency requested on the initial redemption increment.

RedemptionState

Maps an investor and their InvestmentId to the corresponding RedeemState.