Trait pallet_foreign_investments::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen;
    type CurrencyId: Parameter + Member + Copy + TypeInfo + MaxEncodedLen;
    type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug;
    type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo;
    type InvestmentId: TrancheCurrency<Self::PoolId, Self::TrancheId> + Clone + Member + Parameter + Copy + MaxEncodedLen;
    type Investment: InvestmentT<Self::AccountId, Amount = Self::Balance, CurrencyId = Self::CurrencyId, Error = DispatchError, InvestmentId = Self::InvestmentId> + InvestmentCollector<Self::AccountId, Error = DispatchError, InvestmentId = Self::InvestmentId, Result = ()>;
    type BalanceRatio: Parameter + Member + FixedPointNumber + EnsureMul + EnsureDiv + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen;
    type DefaultTokenSellRatio: Get<Self::BalanceRatio>;
    type TokenSwapOrderId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen;
    type TokenSwaps: TokenSwaps<Self::AccountId, CurrencyId = Self::CurrencyId, Balance = Self::Balance, OrderId = Self::TokenSwapOrderId, OrderDetails = Swap<Self::Balance, Self::CurrencyId>, SellRatio = Self::BalanceRatio>;
    type DecreasedForeignInvestOrderHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignDecreaseInvest<Self::Balance, Self::CurrencyId>, Error = DispatchError>;
    type CollectedForeignRedemptionHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignCollect<Self::Balance, Self::CurrencyId>, Error = DispatchError>;
    type CollectedForeignInvestmentHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignCollect<Self::Balance, Self::CurrencyId>, Error = DispatchError>;
    type CurrencyConverter: IdentityCurrencyConversion<Balance = Self::Balance, Currency = Self::CurrencyId, Error = DispatchError>;
    type PoolInspect: PoolInspect<Self::AccountId, Self::CurrencyId, PoolId = Self::PoolId, TrancheId = Self::TrancheId>;

}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

Type representing the weight of this pallet

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen

The source of truth for the balance of accounts

type CurrencyId: Parameter + Member + Copy + TypeInfo + MaxEncodedLen

The currency type of transferrable tokens

type PoolId: Member + Parameter + Default + Copy + HasCompact + MaxEncodedLen + Debug

The pool id type required for the investment identifier

type TrancheId: Member + Parameter + Default + Copy + MaxEncodedLen + TypeInfo

The tranche id type required for the investment identifier

type InvestmentId: TrancheCurrency<Self::PoolId, Self::TrancheId> + Clone + Member + Parameter + Copy + MaxEncodedLen

The investment identifying type required for the investment type

type Investment: InvestmentT<Self::AccountId, Amount = Self::Balance, CurrencyId = Self::CurrencyId, Error = DispatchError, InvestmentId = Self::InvestmentId> + InvestmentCollector<Self::AccountId, Error = DispatchError, InvestmentId = Self::InvestmentId, Result = ()>

The internal investment type which handles the actual investment on top of the wrapper implementation of this Pallet

type BalanceRatio: Parameter + Member + FixedPointNumber + EnsureMul + EnsureDiv + MaybeSerializeDeserialize + TypeInfo + MaxEncodedLen

Type for price ratio for cost of incoming currency relative to outgoing

type DefaultTokenSellRatio: Get<Self::BalanceRatio>

The default sell rate for token swaps which will be applied to all swaps created/updated through Foreign Investments.

Example: Say this rate is set to 3/2, then the incoming currency should never cost more than 1.5 of the outgoing currency.

NOTE: Can be removed once we implement a more sophisticated swap price discovery. For now, this should be set to one.

type TokenSwapOrderId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen

The token swap order identifying type

type TokenSwaps: TokenSwaps<Self::AccountId, CurrencyId = Self::CurrencyId, Balance = Self::Balance, OrderId = Self::TokenSwapOrderId, OrderDetails = Swap<Self::Balance, Self::CurrencyId>, SellRatio = Self::BalanceRatio>

The type which exposes token swap order functionality such as placing and cancelling orders

type DecreasedForeignInvestOrderHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignDecreaseInvest<Self::Balance, Self::CurrencyId>, Error = DispatchError>

The hook type which acts upon a finalized investment decrement.

type CollectedForeignRedemptionHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignCollect<Self::Balance, Self::CurrencyId>, Error = DispatchError>

The hook type which acts upon a finalized redemption collection.

type CollectedForeignInvestmentHook: StatusNotificationHook<Id = ForeignInvestmentInfo<Self::AccountId, Self::InvestmentId, ()>, Status = ExecutedForeignCollect<Self::Balance, Self::CurrencyId>, Error = DispatchError>

The hook type which acts upon a finalized redemption collection.

type CurrencyConverter: IdentityCurrencyConversion<Balance = Self::Balance, Currency = Self::CurrencyId, Error = DispatchError>

Type which provides a conversion from one currency amount to another currency amount.

NOTE: Restricting to IdentityCurrencyConversion is solely a short-term MVP solution. In the near future, this type must be restricted to a more sophisticated trait which provides unidirectional conversions based on an oracle, dynamic prices or at least conversion ratios based on specific currency pairs.

type PoolInspect: PoolInspect<Self::AccountId, Self::CurrencyId, PoolId = Self::PoolId, TrancheId = Self::TrancheId>

The source of truth for pool currencies.