Module pallet_foreign_investments::types

Enums

InvestState

Reflects all states a foreign investment can have until it is processed as an investment via <T as Config>::Investment. This includes swapping it into a pool currency or back, if the investment is decreased before it is fully processed.

InvestTransition

Reflects all state transitions of an InvestmentState which can be externally triggered, i.e. by (partially) fulfilling a token swap order or updating an unprocessed investment.

RedeemState

Reflects all states a foreign redemption can have until transferred to the corresponding source domain.

RedeemTransition

Reflects all state transitions of a RedeemState which can be externally triggered, i.e. by (partially) fulfilling a token swap order or updating an unprocessed redemption.

TokenSwapReason

Reflects the reason for the last token swap update such that it can be updated accordingly if the last and current reason mismatch.

Traits

InvestStateConfig

Restriction of pallet_foreign_investments::Config trait to support currency conversion in the InvestState.